--- a/src/sonic-config-engine/setup.py
+++ b/src/sonic-config-engine/setup.py
@@ -40,7 +40,6 @@ else:
 # Common modules for python2 and python3
 py_modules = [
     'config_samples',
-    'lazy_re',
     'minigraph',
     'openconfig_acl',
     'portconfig',